Fix issue where /exec/init calls put kubeconfig in first container
When /exec/init is called (e.g. in the web terminal operator), it
resolves the first container that supports a shell, but does not use
this container for injecting kubeconfig. When no container name is
provided, the kubeconfig is injected into the first container, which
causes kubeconfig injection to fail when the first container in the pod
does not support a shell.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>