Fix issue creating kubeconfig for users with numeric names
When creating a kubeconfig in the tooling container on an /exec/init
call, we were using

  echo "<config>" > $KUBECONFIG

Using double-quotes in the echo resulted in stripping all double quotes
out of the actual config string. Normally this would not cause an issue
as alphanumeric strings are parsed as strings in yaml even without
quotes, but for users whose names contain only numbers (or are formatted
to be parsed as numbers, e.g. 0xbeef), this meant that the kubeconfig
was invalid as it expects a string for username instead of a number.

Instead we cleanup the formatting for that exec command and instead
enclose it in single quotes, which are unused by the yaml marshaller

Signed-off-by: Angel Misevski <amisevsk@redhat.com>