Support kube:admin as authenticated user (UID="")
The default kubeadmin user on OpenShift clusters does not have a UID and
so gets an empty creator ID. To accomodate this while still requiring
authenticated user ID to be specified, set the default value for
authenticated user ID to the null string terminator (\x00) and check
against this default value.

Go's default flags package does not allow specifying flags as mandatory.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>